+++ src/OpenSimCreator/Documents/MeshWarper/TPSDocumentCSVFlags.hpp
@@ -4,19 +4,19 @@
 
 namespace osc
 {
-	enum class TPSDocumentCSVFlags {
-		None     = 0,
-		NoHeader = 1<<0,
-		NoNames  = 1<<1,
-	};
+    enum class TPSDocumentCSVFlags {
+        None     = 0,
+        NoHeader = 1<<0,
+        NoNames  = 1<<1,
+    };
 
-	constexpr TPSDocumentCSVFlags operator|(TPSDocumentCSVFlags lhs, TPSDocumentCSVFlags rhs)
-	{
-		return static_cast<TPSDocumentCSVFlags>(osc::to_underlying(lhs) | osc::to_underlying(rhs));
-	}
+    constexpr TPSDocumentCSVFlags operator|(TPSDocumentCSVFlags lhs, TPSDocumentCSVFlags rhs)
+    {
+        return static_cast<TPSDocumentCSVFlags>(osc::to_underlying(lhs) | osc::to_underlying(rhs));
+    }
 
-	constexpr bool operator&(TPSDocumentCSVFlags lhs, TPSDocumentCSVFlags rhs)
-	{
-		return (osc::to_underlying(lhs) & osc::to_underlying(rhs)) != 0;
-	}
+    constexpr bool operator&(TPSDocumentCSVFlags lhs, TPSDocumentCSVFlags rhs)
+    {
+        return (osc::to_underlying(lhs) & osc::to_underlying(rhs)) != 0;
+    }
 }